Correct Asymptotic
* ``__str__`` prints even the very small coefficients. It is not its role to decide what coefficients are negligible in the context.
* ``limit`` does not use closeness to 0 either. It is not its role to decide what coefficients are negligible in the context.
Motiviation: in both cases, it is more the role of functions like __mult__ to judge for example that ``self.mu - other.mu`` can be considered null because ``self.mu`` and ``other.mu`` are very close. This is sounder than checking if ``self.mu - other.mu`` is close to zero (with which tolerance?).