Implement the "offset method with trio approximation correction"
In some very rare cases, both pivots are difficult but due to the numerical approximation in the trio event, the pseudo-offsets psi_i or psi_k may be greater than 1, which prevents from using the offset method (infinite sums). So we consider that the pseudo-offset is lower than 1 and extremely close to 1.