Fix a bug due to numerical optimization
Motivation: In very rare cases, the optimizer give a completely stupid result: in the tau mentionned in test_event_trio, it used to lead to phi_b much greater than 1, whereas we know theoretically that phi_b < 1. This leads to many problems when computing the weak pivot, the best response... that all rely on the condition phi_b < 1 in that case.
Solution: based on the knowledge of easy/difficult pivots, we provide bounds to the optimizer that ensure that phi_a, phi_b and phi_c are lower or greater than 1 when they must be.